Repotting and docs.
