+++ primitives/src/ed25519.rs
@@ -4,6 +4,7 @@ use untrusted;
 use ring::{rand, signature};
 use rustc_hex::FromHex;
 
+/// Verify a message without type checking the parameters' types for the right size.
 pub fn verify(sig: &[u8], message: &[u8], public: &[u8]) -> bool {
 	let public_key = untrusted::Input::from(public);
 	let msg = untrusted::Input::from(message);
@@ -27,9 +28,12 @@ pub struct Pair(signature::Ed25519KeyPair);
 pub struct Signature ([u8; 64]);
 
 impl Signature {
+	/// A new signature from the given 64-byte `data`.
 	pub fn from(data: [u8; 64]) -> Self {
 		Signature(data)
 	}
+
+	/// A new signature from the given slice that should be 64 bytes long.
 	pub fn from_slice(data: &[u8]) -> Self {
 		let mut r = [0u8; 64];
 		r.copy_from_slice(data);
@@ -50,9 +54,12 @@ impl AsRef<[u8]> for Signature {
 }
 
 impl Public {
+	/// A new instance from the given 32-byte `data`.
 	pub fn from(data: [u8; 32]) -> Self {
 		Public(data)
 	}
+
+	/// A new instance from the given slice that should be 32 bytes long.
 	pub fn from_slice(data: &[u8]) -> Self {
 		let mut r = [0u8; 32];
 		r.copy_from_slice(data);
@@ -166,53 +173,9 @@ impl PartialEq for Signature {
 	}
 }
 
-pub mod hexdisplay {
-
-	pub struct HexDisplay<'a>(&'a [u8]);
-
-	impl<'a> HexDisplay<'a> {
-		pub fn from(d: &'a AsBytesRef) -> Self { HexDisplay(d.as_bytes_ref()) }
-	}
-
-	impl<'a> ::std::fmt::Display for HexDisplay<'a> {
-		fn fmt(&self, fmtr: &mut ::std::fmt::Formatter) -> Result<(), ::std::fmt::Error> {
-			for byte in self.0 {
-				try!( fmtr.write_fmt(format_args!("{:02x}", byte)));
-			}
-			Ok(())
-		}
-	}
-
-	pub trait AsBytesRef {
-		fn as_bytes_ref(&self) -> &[u8];
-	}
-
-	impl AsBytesRef for [u8] {
-		fn as_bytes_ref(&self) -> &[u8] { &self }
-	}
-
-	impl AsBytesRef for Vec<u8> {
-		fn as_bytes_ref(&self) -> &[u8] { &self }
-	}
-
-	macro_rules! impl_non_endians {
-		( $( $t:ty ),* ) => { $(
-			impl AsBytesRef for $t {
-				fn as_bytes_ref(&self) -> &[u8] { &self[..] }
-			}
-		)* }
-	}
-
-	impl_non_endians!([u8; 1], [u8; 2], [u8; 3], [u8; 4], [u8; 5], [u8; 6], [u8; 7], [u8; 8],
-		[u8; 10], [u8; 12], [u8; 14], [u8; 16], [u8; 20], [u8; 24], [u8; 28], [u8; 32], [u8; 40],
-		[u8; 48], [u8; 56], [u8; 64], [u8; 80], [u8; 96], [u8; 112], [u8; 128]);
-
-}
-
 #[cfg(test)]
 mod test {
 	use super::*;
-	use super::hexdisplay::HexDisplay;
 
 	#[test]
 	fn test_vector_should_work() {
@@ -238,7 +201,6 @@ mod test {
 	fn seeded_pair_should_work() {
 		let pair = Pair::from_seed(b"12345678901234567890123456789012");
 		let public = pair.public();
-		println!("{}", HexDisplay::from(&public.0));
 		assert_eq!(public, "2f8c6129d816cf51c374bc7f08c3e63ed156cf78aefb4a6550d97b87997977ee".into());
 		let message = b"Something important";
 		let signature = pair.sign(&message[..]);
@@ -252,8 +214,6 @@ mod test {
 		assert_eq!(public, "2f8c6129d816cf51c374bc7f08c3e63ed156cf78aefb4a6550d97b87997977ee".into());
 		let message = FromHex::from_hex("2f8c6129d816cf51c374bc7f08c3e63ed156cf78aefb4a6550d97b87997977ee00000000000000000228000000d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511a4500000000000000").unwrap();
 		let signature = pair.sign(&message[..]);
-		println!("pub: {}", HexDisplay::from(&public.0));
-		println!("sig: {}", HexDisplay::from(&signature.0));
 		assert!(signature.verify(&message[..], &public));
 	}
 }
